Deprecate my own silly crypto
